requirements.txt
requests==2.26.0
Django==3.2.8
djangorestframework==3.12.4
PyJWT==2.3.0
